requirements/base.txt
anyio==3.6.2
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
click==8.1.3
defusedxml==0.7.1
diff-match-patch==20200713
Django
django-debug-toolbar==3.2.2
django-import-export==3.0.2
et-xmlfile==1.1.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
markdown2==2.4.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
psycopg2-binary==2.9.2
python-telegram-bot==20.0
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.2
tablib==3.3.0
URLObject==2.4.3
uvicorn==0.20.0
xlrd==2.0.1
xlwt==1.3.0
